+++ src/sys/xml/xml_obj.cpp
@@ -11,22 +11,23 @@
 
 namespace shawn
 {
-	namespace xml 
+	namespace xml
 	{
 
 		XMLObj::XMLObj():xml_data_(0) {
 			current_line_read_ = -1;
 		}
-		
+
 		// ----------------------------------------------------------------------
 		XMLObj::~XMLObj() {
-			// de-allocate memory 
+			// de-allocate memory
 			xml_data_.clear();
 		}
-		
+
 
 		bool XMLObj::eof() {
-			if ((current_line_read_<0) || (current_line_read_ < xml_data_.size())) {
+			// cast current_line_read_ to avoid implicit integer conversion
+			if ((current_line_read_<0) || (static_cast<unsigned long>(current_line_read_) < xml_data_.size())) {
 				return false;
 			} else {
 				return true;
@@ -37,7 +38,7 @@ namespace shawn
 		void XMLObj::read(char * buf, int len ) {
 			current_line_read_ = current_line_read_+1;
 			std::string temp = *(xml_data_.end()-(xml_data_.size()-current_line_read_));
-			for (int i=0; i<temp.size(); i++) {
+			for (unsigned int i=0; i<temp.size(); i++) {
 				buf[i] = temp[i];
 			}
 		}
@@ -47,8 +48,8 @@ namespace shawn
 			int gcount = temp.size();
 			return gcount;
 		}
-				
-		void XMLObj::appendLine(std::string line) {	
+
+		void XMLObj::appendLine(std::string line) {
 			xml_data_.push_back(line);
 			// std::cerr << "appendLine " << *(xml_data_.end()-1) << std::endl;
 		}
